% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_params}
\alias{build_params}
\title{Build API Parameter List}
\usage{
build_params(...)
}
\arguments{
\item{...}{Named parameters to include in the API call}
}
\value{
List of non-null parameters
}
\description{
Internal helper function to build parameter lists for API calls.
}
\keyword{internal}
