% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call-ihme-api.R
\name{call_ihme_api}
\alias{call_ihme_api}
\title{Call the IHME SDG API}
\usage{
call_ihme_api(api_url, ...)
}
\arguments{
\item{api_url}{url endpoint}

\item{...}{additional query parameters}
}
\value{
JSON response from the API, usually parsed to a list
}
\description{
Call the IHME SDG API
}
\examples{
\dontshow{if (interactive() && curl::has_internet()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
call_ihme_api("GetLocation")
\dontshow{\}) # examplesIf}
}
