% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-cached-hash-version.R
\name{check_cached_data}
\alias{check_cached_data}
\title{Check if cached data is current}
\usage{
check_cached_data(file)
}
\arguments{
\item{file}{Path to cached data file}
}
\value{
List with cached data if current, NULL if needs refresh
}
\description{
Check if cached data is current
}
