% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_location_lookup}
\alias{create_location_lookup}
\title{Create Location Lookup Table}
\usage{
create_location_lookup()
}
\value{
Data frame with location codes and names for easy lookup
}
\description{
Create Location Lookup Table
}
\examples{
\dontshow{if (interactive() && curl::has_internet()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
create_location_lookup()
\dontshow{\}) # examplesIf}
}
