% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{download_all_ihme_metadata}
\alias{download_all_ihme_metadata}
\title{Download All IHME Metadata}
\usage{
download_all_ihme_metadata(file = "ihme_metadata.rds", force = FALSE)
}
\arguments{
\item{file}{Path to save data (default: "ihme_metadata.rds")}

\item{force}{Force redownload, even if file exists and is current (default: FALSE)}
}
\value{
List containing all IHME metadata
}
\description{
Downloads and caches all available metadata from the IHME API.
Downloads ONLY metadata, not actual SDG indicator data.
}
\examples{
\dontshow{if (interactive() && curl::has_internet()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
metadata <- download_all_ihme_metadata()
metadata <- download_all_ihme_metadata(force = TRUE)
\dontshow{\}) # examplesIf}
}
