% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-metadata.R
\name{get_age_groups}
\alias{get_age_groups}
\title{Get Age Groups}
\usage{
get_age_groups(age_group_id = NULL)
}
\arguments{
\item{age_group_id}{Optional age group ID to filter results}
}
\value{
Data frame of age groups
}
\description{
Get Age Groups
}
\examples{
\dontshow{if (interactive() && curl::has_internet()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_age_groups()
get_age_groups("1")
\dontshow{\}) # examplesIf}
}
