% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{get_all_indicators}
\alias{get_all_indicators}
\title{Get All SDG Indicators}
\usage{
get_all_indicators()
}
\value{
Data frame of all indicators
}
\description{
Get all indicators across all categories.
}
\examples{
\dontshow{if (interactive() && curl::has_internet()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_all_indicators()
\dontshow{\}) # examplesIf}
}
