% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{get_all_locations}
\alias{get_all_locations}
\title{Get All Locations}
\usage{
get_all_locations()
}
\value{
Data frame of all locations
}
\description{
Get all available locations (countries and regions).
}
\examples{
\dontshow{if (interactive() && curl::has_internet()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_all_locations()
\dontshow{\}) # examplesIf}
}
