% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call-ihme-api.R
\name{get_api_version}
\alias{get_api_version}
\title{Get API Version}
\usage{
get_api_version()
}
\value{
API version information
}
\description{
Get API Version
}
\examples{
\dontshow{if (interactive() && curl::has_internet()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_api_version()
\dontshow{\}) # examplesIf}
}
