% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-metadata.R
\name{get_goals}
\alias{get_goals}
\title{Get SDG Goals}
\usage{
get_goals(goal_id = NULL)
}
\arguments{
\item{goal_id}{Optional vector of goal IDs to filter results}
}
\value{
Data frame of SDG goals
}
\description{
Get SDG Goals
}
\examples{
\dontshow{if (interactive() && curl::has_internet()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_goals()
get_goals(c(1, 3))
\dontshow{\}) # examplesIf}
}
