% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-metadata.R
\name{get_indicators}
\alias{get_indicators}
\title{Get SDG Indicators}
\usage{
get_indicators(indicator_id = NULL)
}
\arguments{
\item{indicator_id}{Optional indicator ID to filter results}
}
\value{
Data frame of indicators
}
\description{
Get SDG Indicators
}
\examples{
\dontshow{if (interactive() && curl::has_internet()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_indicators()
get_indicators("1.1.1")
\dontshow{\}) # examplesIf}
}
