% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-metadata.R
\name{get_locations}
\alias{get_locations}
\title{Get Locations}
\usage{
get_locations(location_id = NULL)
}
\arguments{
\item{location_id}{Optional location ID to filter results}
}
\value{
Data frame of locations (countries, regions, etc.)
}
\description{
Get Locations
}
\examples{
\dontshow{if (interactive() && curl::has_internet()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_locations()
get_locations("102")
\dontshow{\}) # examplesIf}
}
