% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{get_results_by_location}
\alias{get_results_by_location}
\title{Get Results by Location}
\usage{
get_results_by_location(
  location_id = NULL,
  indicator_id = NULL,
  year = NULL,
  sex_id = NULL,
  age_group_id = NULL,
  scenario = NULL
)
}
\arguments{
\item{location_id}{Location ID to retrieve results for}

\item{indicator_id}{Optional vector of indicator IDs to filter results}

\item{year}{Optional vector of years (1990-2030) to filter results}

\item{sex_id}{Optional vector of sex IDs to filter results}

\item{age_group_id}{Optional vector of age group IDs to filter results}

\item{scenario}{Optional vector of scenario IDs to filter results}
}
\value{
Data frame of results for the specified location
}
\description{
Get Results by Location
}
\examples{
\dontshow{if (interactive() && curl::has_internet()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_results_by_location(location_id = 102, indicator_id = c(1, 2), year = c(2020, 2021))
\dontshow{\}) # examplesIf}
}
