% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{get_results_by_target}
\alias{get_results_by_target}
\title{Get Results by Target}
\usage{
get_results_by_target(
  target_id = NULL,
  location_id = NULL,
  year = NULL,
  sex_id = NULL,
  age_group_id = NULL,
  scenario = NULL
)
}
\arguments{
\item{target_id}{Target ID to retrieve results for}

\item{location_id}{Optional vector of location IDs to filter results}

\item{year}{Optional vector of years (1990-2030) to filter results}

\item{sex_id}{Optional vector of sex IDs to filter results}

\item{age_group_id}{Optional vector of age group IDs to filter results}

\item{scenario}{Optional vector of scenario IDs to filter results}
}
\value{
Data frame of results for the specified target
}
\description{
Get Results by Target
}
\examples{
\dontshow{if (interactive() && curl::has_internet()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_results_by_target(target_id = 1, location_id = c(102, 6), year = c(2020, 2021))
\dontshow{\}) # examplesIf}
}
