% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{get_results_by_year}
\alias{get_results_by_year}
\title{Get Results by Year}
\usage{
get_results_by_year(
  year = NULL,
  indicator_id = NULL,
  location_id = NULL,
  sex_id = NULL,
  age_group_id = NULL,
  scenario = NULL
)
}
\arguments{
\item{year}{Year (1990-2030) to retrieve results for}

\item{indicator_id}{Optional vector of indicator IDs to filter results}

\item{location_id}{Optional vector of location IDs to filter results}

\item{sex_id}{Optional vector of sex IDs to filter results}

\item{age_group_id}{Optional vector of age group IDs to filter results}

\item{scenario}{Optional vector of scenario IDs to filter results}
}
\value{
Data frame of results for the specified year
}
\description{
Get Results by Year
}
\examples{
\dontshow{if (interactive() && curl::has_internet()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_results_by_year(year = 2020, indicator_id = c(1, 2), location_id = c(102, 6))
\dontshow{\}) # examplesIf}
}
