% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-metadata.R
\name{get_scenarios}
\alias{get_scenarios}
\title{Get Scenarios}
\usage{
get_scenarios(scenario = NULL)
}
\arguments{
\item{scenario}{Optional scenario to filter results}
}
\value{
Data frame of projection scenarios
}
\description{
Get Scenarios
}
\examples{
\dontshow{if (interactive() && curl::has_internet()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_scenarios()
get_scenarios("reference")
\dontshow{\}) # examplesIf}
}
