% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-metadata.R
\name{get_sex_categories}
\alias{get_sex_categories}
\title{Get Sex Categories}
\usage{
get_sex_categories(sex_id = NULL)
}
\arguments{
\item{sex_id}{Optional sex ID to filter results}
}
\value{
Data frame of sex categories
}
\description{
Get Sex Categories
}
\examples{
\dontshow{if (interactive() && curl::has_internet()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_sex_categories()
get_sex_categories("1")
\dontshow{\}) # examplesIf}
}
