% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-metadata.R
\name{get_targets}
\alias{get_targets}
\title{Get SDG Targets}
\usage{
get_targets(target_id = NULL)
}
\arguments{
\item{target_id}{Optional target ID to filter results}
}
\value{
Data frame of SDG targets
}
\description{
Get SDG Targets
}
\examples{
\dontshow{if (interactive() && curl::has_internet()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_targets()
get_targets(1)
\dontshow{\}) # examplesIf}
}
