% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readihme-package.R
\docType{package}
\name{readihme-package}
\alias{readihme}
\alias{readihme-package}
\title{readihme: Access IHME SDG Data Through R}
\description{
Provides easy access to the Institute for Health Metrics and Evaluation (IHME) Sustainable Development Goals (SDG) API. Functions allow users to retrieve SDG indicator data, metadata, and perform common data analysis tasks with IHME data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/yourusername/readihme}
  \item Report bugs at \url{https://github.com/yourusername/readihme/issues}
}

}
\author{
\strong{Maintainer}: Beth Firipis \email{bfiripis@gmail.com} (\href{https://orcid.org/0009-0008-3477-5003}{ORCID})

}
\keyword{internal}
