% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_years}
\alias{validate_years}
\title{Validate Year Range}
\usage{
validate_years(years)
}
\arguments{
\item{years}{Vector of years to validate}
}
\value{
Logical indicating if all years are valid
}
\description{
Internal function to validate that years are within the acceptable range.
}
\keyword{internal}
